Generalised the Bindingprovider System
